requirements.txt
streamlit
transformers
torch
pillow
gtts
torchvision
matplotlib
scikit-image
scikit-learn
torchvision
nltk
setuptools>=68.2.2
torch
sacremoses
sentencepiece
transformers
sacrebleu
wikipedia-api
googlesearch-python
